# Conflicts:
#	out/production/116Project/Characters/Character.class
#	out/production/116Project/Characters/Enemy.class
#	out/production/116Project/Characters/Fighter.class
#	out/production/116Project/Characters/Healer.class
#	out/production/116Project/Characters/Tank.class
#	out/production/116Project/GameThings/Game.class
#	out/production/116Project/GameThings/Level.class
#	out/production/116Project/Items/Items.class
#	out/production/116Project/Items/Shield.class
#	out/production/116Project/Items/Sword.class
#	out/production/116Project/Items/Wand.class
#	out/production/116Project/Items/Weapon.class
#	src/GameThings/Game.java

+++ src/Items/Wand.java
@@ -2,7 +2,7 @@ package Items;
 
 public class Wand extends Weapon {
 
-    public Wand(String name, int weight, int value, int damage) {
+    public Wand(String name, int weight, int value,double damage) {
         super(name, weight, value, damage);
     }
